requirements.txt
django
google
google-generativeai
python-dotenv
google-auth-oauthlib
md-to-html
markdown2
django-markdownify
whitenoise
plaid-python